+++ src/sll/include/sll/sandbox.h
@@ -163,6 +163,18 @@
 
 
 /**
+ * \flags macro var
+ * \name SLL_SANDBOX_FLAG_DISABLE_THREADS
+ * \group sandbox
+ * \subgroup sandbox-flag
+ * \desc Docs!
+ * \type sll_sandbox_flag_t
+ */
+#define SLL_SANDBOX_FLAG_DISABLE_THREADS 12
+
+
+
+/**
  * \flags check_output func
  * \name sll_get_sandbox_flag
  * \group sandbox