+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Implemented `weakref.sll`, `weakref$weakref_type`, `weakref$NO_OBJECT`, `weakref$get` and `weakref$ref` ([#257])
 - Object creation code to build an array from a pointer and a length: `L`
 - Object creation code to build an unsigned 32-bit integer: `u`
+- Sandbox flag for disabling thread API: `no-thread`
 
 #### Built-in Audit Events